fix(feature-flagging): preserve shard field compatibility

+++ products/feature-flagging/feature-flagging-bootstrap/src/main/java/datadog/trace/api/featureflag/ufc/v1/Shard.java
@@ -5,9 +5,9 @@ import java.util.List;
 public class Shard {
   public final String salt;
   public final List<ShardRange> ranges;
-  public final long totalShards;
+  public final int totalShards;
 
-  public Shard(final String salt, final List<ShardRange> ranges, final long totalShards) {
+  public Shard(final String salt, final List<ShardRange> ranges, final int totalShards) {
     this.salt = salt;
     this.ranges = ranges;
     this.totalShards = totalShards;
